+++ CHANGELOG.md
@@ -68,6 +68,8 @@ release. They will however *never* happen in a patch release.
 }
 ```
 
+* fix a bug where vsmtp crash in case of an ill-formatted email address in 'rcpt' and 'mail from' command's arguments.
+
 ## [2.2.1] - 2023-03-31
 
 ### Added